scripts/requirements.txt
addict
google-api-python-client
google-cloud-storage
httplib2
numpy
pip
pyyaml
requests
